% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acc.retrieve.R
\name{acc.retrieve}
\alias{acc.retrieve}
\title{Retrieve accession numbers and titles for a genbank search}
\usage{
acc.retrieve(
  organism,
  acc.num = FALSE,
  gene = NULL,
  speciesLevel = FALSE,
  npar = 2,
  nSearchesBatch = 499
)
}
\arguments{
\item{organism}{The name of a single taxon (character).}

\item{acc.num}{Logical indicating whether the \code{organism}
is actually a vector or accession numbers.}

\item{gene}{The name of a single gene region (character; optional).}

\item{speciesLevel}{Whether the result should be a species-level dataset (logical).}

\item{npar}{Number of parallel searches (the default is probably the best option).}

\item{nSearchesBatch}{Number of searches per batch}
}
\value{
This function returns an object of class \code{data.frame} that
includes the following columns. First, \code{Species} with the
species name listed in GenBank. Second, \code{Ti} including the title
of the accession in GenBank. Third, \code{Acc} listing the accession
number. Fourth, \code{gene} including the name of the target gene.
}
\description{
Retrieve accession numbers and titles for searches in genbank. This function
is useful for exploring the gene-, population-, and species-level sampling
in genbank.
}
\examples{
\dontrun{
acc.retrieve(organism="Vulpes", gene = 'cytb', species=TRUE)
}
}
\keyword{internal}
