% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getListConstraints.R
\name{getListConstraints}
\alias{getListConstraints}
\title{Get a list of constraints given a taxonomic dataset}
\usage{
getListConstraints(dataset, targetColumns, byClades = FALSE)
}
\description{
Internal used to generate tree constraints
}
\keyword{internal}
