% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sq.retrieve.R
\name{sq.retrieve}
\alias{sq.retrieve}
\title{Retrieve sequences from genbank}
\usage{
sq.retrieve(
  clades = NULL,
  species = NULL,
  genes = NULL,
  maxseqs = 1,
  maxlength = 5000
)
}
\arguments{
\item{clades}{A vector listing taxonomic groups of interest (character).}

\item{species}{A vector listing additional species interest (character).
This argument can be used to define additional target species
in the ingroup or species to be sampled in the outgroup (character).}

\item{genes}{A vector listing gene names of interest (character).}

\item{maxseqs}{Maximum number of sequences to retrieve per search (taxa + gene) (numeric).}

\item{maxlength}{Maximum lenght of the gene sequence (numeric).}
}
\value{
None
}
\description{
Downloads sequences from genbank (nucleotide database) for particular taxa and genes
into a folder called \code{"0.Sequences"}.
}
\examples{
\dontrun{
sq.retrieve(
            clades = c('Felis', 'Vulpes', 'Phoca'),
            species = 'Manis_pentadactyla' ,
            genes = c("ADORA3")
           )
}
}
