% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SW.phruta.R
\docType{data}
\name{SW.phruta}
\alias{SW.phruta}
\title{Scholl and Wiens (2016) phylogenies}
\format{
An object of class \code{"multiphylo"}.
}
\source{
\href{http://dx.doi.org/10.1098/rspb.2016.1334}{Publication}
}
\usage{
data(SW.phruta)
}
\description{
Time-calibrated trees from Scholl and Wiens (2016).
phruta uses them to calibrate resulting phylogenies.
}
\references{
Scholl, Joshua P., and John J. Wiens.
"Diversification rates and species richness
across the Tree of Life." Proceedings of the Royal
Society B: Biological Sciences 283.1838 (2016): 20161334.
}
\keyword{datasets}
