% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.dating.R
\name{tree.dating}
\alias{tree.dating}
\title{Tree dating under treePL or}
\usage{
tree.dating(
  taxonomyFolder = "1.CuratedSequences",
  phylogenyFolder = "3.Phylogeny",
  ...
)
}
\arguments{
\item{taxonomyFolder}{Name of the folder where \code{"1.Taxonomy.csv"},
created duing the
\code{"sq.curate"} step, is stored (character).}

\item{phylogenyFolder}{Name of the folder where
\code{"RAxML_bipartitions.phruta"}, created duing the
\code{"tree.raxml"} step, is stored (character).}

\item{...}{Arguments passed to \code{"geiger::congruify.phylo"}.}
}
\value{
None
}
\description{
Performs tree dating under \code{"treePL"} or \code{"PATHd-8"}
based on secondary calibrations.
Note that \code{"treePL"} or \code{"PATHd-8"} must be installed in your PATH.
How to install \code{"PATHd-8"} in mac
\code{"https://gist.github.com/cromanpa94/a43bc710a17220f71d796d6590ea7fe4"}
and \code{"treePL"} can be installed using homebrew
(brew install brewsci/bio/treepl). Thanks to Brian O'Meara and Jonathan
Chang, respectively.
}
\examples{
\dontrun{
sq.retrieve.direct(
  clades = c("Felis", "Vulpes", "Phoca"),
  species = "Manis_pentadactyla",
  genes = c("ADORA3", "CYTB")
)
sq.curate(
  filterTaxonomicCriteria = "Felis|Vulpes|Phoca|Manis",
  kingdom = "animals", folder = "0.Sequences"
)
sq.aln(folder = "1.CuratedSequences")
tree.raxml(
  folder = "2.Alignments", FilePatterns = "Masked",
  raxml_exec = "raxmlHPC", Bootstrap = 100,
  outgroup = "Manis_pentadactyla"
)
tree.dating(
  taxonomyFolder = "1.CuratedSequences",
  phylogenyFolder = "3.Phylogeny", scale = "treePL"
)
}
}
