% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.roguetaxa.R
\name{tree.roguetaxa}
\alias{tree.roguetaxa}
\title{RogueNaRok within phruta}
\usage{
tree.roguetaxa(folder = "3.Phylogeny", ...)
}
\arguments{
\item{folder}{Name of the folder where the sequences to align are
stored (character).}

\item{...}{Arguments passed to \code{"Rogue::RogueTaxa"}.}
}
\value{
None
}
\description{
Implements the RogueNaRok algorithm for rogue taxon identification
within phruta
}
\examples{
\dontrun{
sq.retrieve.direct(
  clades = c("Felis", "Vulpes", "Phoca"),
  species = "Manis_pentadactyla",
  genes = c("ADORA3", "CYTB")
)
sq.curate(
  filterTaxonomicCriteria = "Felis|Vulpes|Phoca|Manis",
  kingdom = "animals", folder = "0.Sequences"
)
sq.aln(folder = "1.CuratedSequences")
tree.raxml(
  folder = "2.Alignments", FilePatterns = "Masked",
  raxml_exec = "raxmlHPC", Bootstrap = 100,
  outgroup = "Manis_pentadactyla"
)
tree.roguetaxa(folder = "3.Phylogeny")
}
}
