% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acc.retrieve.R
\name{acc.retrieve}
\alias{acc.retrieve}
\title{Retrieve accession numbers and titles for a given search in genbank}
\usage{
acc.retrieve(
  organism,
  acc.num = FALSE,
  gene = NULL,
  speciesLevel = FALSE,
  npar = 2
)
}
\arguments{
\item{organism}{The name of a single taxon (character).}

\item{gene}{The name of a single gene region (character; optional).}

\item{speciesLevel}{Whether the result should be a species-level dataset (logical).}
}
\value{
data.frame
}
\description{
Retrieve accession numbers and titles for searches in genbank. This function
is useful for exploring the gene-, population-, and species-level sampling
in genbank.
}
\examples{
\dontrun{
acc.retrieve(organism="Vulpes", gene = 'cytb', species=TRUE)
}
}
