% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sq.aln.R
\name{sq.aln}
\alias{sq.aln}
\title{Align sequences}
\usage{
sq.aln(folder = "1.CuratedSequences", FilePatterns = "renamed", mask = T, ...)
}
\arguments{
\item{folder}{Name of the folder where the sequences to align are stored (character).}

\item{FilePatterns}{A string that is common to all the target files in the relevant folder (character).}

\item{mask}{Removes ambiguous sites (TRUE or FALSE).}

\item{...}{Arguments passed to \code{"DECIPHER::AlignSeqs"}.}
}
\value{
None
}
\description{
Perform multiple sequence alignment on all the fasta files
saved in a given folder. Alignment is conducted using
\code{"BiocManager::install("DECIPHER")"}. Note that this function first
optimizes the direction of the sequences, then aligns using \code{"BDECIPHER"},
and finally masks the resulting alignment (optionally but does it per default). The
masking step includes removing common gaps across all the species and removing
highly ambiguous positions. The resulting aligned sequences are stored to a new folder
\code{"2.Alignments"}.'
}
\examples{
\dontrun{
sq.retrieve(
            clades = c('Felis', 'Vulpes', 'Phoca'),
            species = 'Manis_pentadactyla' ,
            genes = c("ADORA3")
           )
sq.curate(filterTaxonomicCriteria='Felis|Vulpes|Phoca|Manis',
          kingdom='animals', folder='0.Sequences')
sq.aln(folder='1.CuratedSequences')
}
}
