% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sq.curate.R
\name{sq.curate}
\alias{sq.curate}
\title{Curate sequences from genbank}
\usage{
sq.curate(
  filterTaxonomicCriteria = NULL,
  mergeGeneFiles = NULL,
  database = "gbif",
  kingdom = NULL,
  folder = "0.Sequences",
  removeOutliers = TRUE,
  minSeqs = 5,
  threshold = 0.05,
  ranks = c("kingdom", "phylum", "class", "order", "family", "genus", "species")
)
}
\arguments{
\item{filterTaxonomicCriteria}{A single string of terms (delimited using "|")
listing all the strings that could be used to
identify the species that should
be in the dataset (character).}

\item{mergeGeneFiles}{A named list, with each element being a character vector
indicating the names of the files in \code{"0.Sequences"}
that need to be combined into a single fasta file. For instance,
you can use this argument to combine CO1 and COI.}

\item{database}{A name of a database with taxonomic information.
Although 'gbif' is faster, it only has information for
animals and plants. Other databases follow
taxize::classification.}

\item{kingdom}{Optional and only used when database='gbif'.
Two possible options: "animals" or "plants."}

\item{folder}{The name of the folder where the original sequences are
located (character).}

\item{removeOutliers}{Whether  \code{odseq:odseq} should be used to remove outliers}

\item{minSeqs}{minimum number of sequences per locus}

\item{threshold}{Relative to \code{odseq::odseq}. Only important if
\code{removeOutliers = TRUE}}

\item{ranks}{The taxonomic ranks used to examine the taxonomy of the species
in the \code{0.Sequences} folder.}
}
\value{
None
}
\description{
After downloading sequences from genbank, this function
curates sequences based on taxonomic
information. Note that this function provides two summary datasets.
First, the accession numbers.
Second, the taxonomic information for each species in the database.
The taxonomy strictly follows
the gbif taxonomic backbone. The resulting files are saved
to \code{"1.CuratedSequences"}. The
resulting files also have the most recent curated taxonomy
following the gbif (or selected database) taxonomic backbone.
}
\examples{
\dontrun{
sq.retrieve.direct(
  clades = c("Felis", "Vulpes", "Phoca"),
  species = "Manis_pentadactyla",
  genes = c("ADORA3", "CYTB")
)
sq.curate(
  filterTaxonomicCriteria = "Felis|Vulpes|Phoca|Manis",
  database = "gbif", kingdom = "animals",
  folder = "0.Sequences"
)
}
}
