% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sq.partitionfinder1.R
\name{sq.partitionfinderv1}
\alias{sq.partitionfinderv1}
\title{Run Partitionfinder v.1}
\usage{
sq.partitionfinderv1(
  folderAlignments = "2.Alignments",
  FilePatterns = "Masked",
  folderPartitionFinder = "2.1.PartitionFinderv1",
  models = "all"
)
}
\arguments{
\item{folderAlignments}{Name of the folder where the sequences to align are stored (character).}

\item{FilePatterns}{A string that is common to all the target files in the relevant folder (character). Note that
this argument can be set to \code{"NULL"} if no specific pattern wants to be analyzed.}

\item{folderPartitionFinder}{Name of the new folder where the output files are stored (string).}

\item{models}{Models to run in partitionfinder (string).}
}
\value{
None
}
\description{
This function runs partitionfinder v1 within phruta. For now, all analyses are based on genes.
}
\examples{
\dontrun{
sq.retrieve(
  clades = c("Felis", "Vulpes", "Phoca"),
  species = "Manis_pentadactyla",
  genes = c("ADORA3")
)
sq.curate(
  filterTaxonomicCriteria = "Felis|Vulpes|Phoca|Manis",
  kingdom = "animals", folder = "0.Sequences"
)
sq.aln(folder = "1.CuratedSequences")
sq.partitionfinder1(
  folderAlignments = "2.Alignments",
  FilePatterns = "Masked",
  folderPartitionFinder = "2.1.PartitionFinderv1",
  models = "all"
)
}
}
