% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.raxml.R
\name{tree.raxml}
\alias{tree.raxml}
\title{Tree inference under RAxML}
\usage{
tree.raxml(
  folder = "2.Alignments",
  FilePatterns = "Masked_",
  raxml_exec = "raxmlHPC",
  Bootstrap = 100,
  outgroup,
  partitioned = FALSE,
  ...
)
}
\arguments{
\item{folder}{Name of the folder where the sequences to align are stored
(character).}

\item{FilePatterns}{A string that is common to all the target files
in the relevant folder (character). Note that
this argument can be set to \code{"NULL"} if no specific
pattern wants to be analized.}

\item{raxml_exec}{Where to find \code{"RAxML"} or how to run it from the
console? (string).}

\item{Bootstrap}{Number of bootstrap replicates (numeric).}

\item{outgroup}{A single string of comma-separated tip labels to be
used as outgroup in \code{"RAxML"} See \code{"RAxML"}
documentation for more details (character).}

\item{partitioned}{Whether analyses should be partitioned by gene (Logical).}

\item{...}{Arguments passed to \code{"ips::raxml"}.}
}
\value{
None
}
\description{
Performs tree inference under \code{"RAxML"} for aligned fasta sequences in
a given folder (default is \code{"2.Alignments"}). Note that you need at
least two gene regions to run a partitioned analysis.
}
\examples{
\dontrun{
sq.retrieve.direct(
  clades = c("Felis", "Vulpes", "Phoca"),
  species = "Manis_pentadactyla",
  genes = c("ADORA3", "CYTB")
)
sq.curate(
  filterTaxonomicCriteria = "Felis|Vulpes|Phoca|Manis",
  kingdom = "animals", folder = "0.Sequences"
)
sq.aln(folder = "1.CuratedSequences")
tree.raxml(
  folder = "2.Alignments", FilePatterns = "Masked",
  raxml_exec = "raxmlHPC", Bootstrap = 100,
  outgroup = "Manis_pentadactyla"
)
}
}
